requirements.txt
requests
azure-identity
azure-mgmt-compute